backend/requirements.txt
fastapi
uvicorn
sqlalchemy[asyncio]
asyncpg
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
